change version type

+++ api/cashu/types.go
@@ -192,7 +192,7 @@ func (keyset *MintKey) GetPubKey() *secp256k1.PublicKey {
 type Seed struct {
 	Active      bool
 	CreatedAt   int64
-	Version     int
+	Version     uint32
 	Unit        string
 	Id          string
 	InputFeePpk uint `json:"input_fee_ppk" db:"input_fee_ppk"`
@@ -469,7 +469,7 @@ type BasicKeysetResponse struct {
 	Unit        string `json:"unit"`
 	Active      bool   `json:"active"`
 	InputFeePpk uint   `json:"input_fee_ppk"`
-	Version     uint64
+	Version     uint32
 }
 
 type PostCheckStateRequest struct {
